Move Applicative[Gen] to gem.arb package object
